src/geminiai_generation.py
# import os
#
from src.config import cfg
# from src.logger import logger
import google.generativeai as genai

"""
At the command line, only need to run once to install the package via pip:

$ pip install google-generativeai
"""


def func1():
    genai.configure(api_key=cfg.GEMINIAI_API_KEY)
    # Set up the model
    generation_config = {
        "temperature": 0.9,
        "top_p": 1,
        "top_k": 1,
        "max_output_tokens": 2048,
    }
    safety_settings = [
        {
            "category": "HARM_CATEGORY_HARASSMENT",
            "threshold": "BLOCK_MEDIUM_AND_ABOVE"
        },
        {
            "category": "HARM_CATEGORY_HATE_SPEECH",
            "threshold": "BLOCK_MEDIUM_AND_ABOVE"
        },
        {
            "category": "HARM_CATEGORY_SEXUALLY_EXPLICIT",
            "threshold": "BLOCK_MEDIUM_AND_ABOVE"
        },
        {
            "category": "HARM_CATEGORY_DANGEROUS_CONTENT",
            "threshold": "BLOCK_MEDIUM_AND_ABOVE"
        },
    ]

    model = genai.GenerativeModel(model_name="gemini-1.0-pro",
                                  generation_config=generation_config,
                                  safety_settings=safety_settings)
    convo = model.start_chat(history=[
    ])
    return convo




def run_geminiai_generation(input_data: str, prompt: str):
    convo=func1()
    convo.send_message(input_data)
    return convo.last.text

