requirements.txt
openai==0.27.8
moviepy==1.0.3
numpy==1.25.0
PyYAML==6.0
nltk==3.8.1

soundfile==0.12.1
pytube==15.0.0

pytest==7.4.0
black==23.3.0
isort==5.12.0
